[enh #17] Drawerの排除とインターフェース関連の実装
