--- LibraryProject/SharokuLibrary/sl/Drawer/dx11Drawer/dx11Drawer.cpp
@@ -1,27 +1,0 @@
-﻿//==================================================================================================================================//
-//!< @file		dx11Drawer.cpp
-//!< @brief		dx11::Drawerクラス実装
-//!< @author	T.Haga
-//==================================================================================================================================//
-
-
-#include "dx11Drawer.h"
-
-
-namespace sl
-{
-namespace dx11
-{
-
-Drawer::Drawer(void)
-{}
-
-Drawer::~Drawer(void)
-{}
-
-}	// namespace dx11
-}	// namespace sl
-
-//==================================================================================================================================//
-// END OF FILE
-//==================================================================================================================================//
