--- LibraryProject/SharokuLibrary/sl/Drawer/dx11Drawer/dx11Drawer.h
@@ -1,36 +1,0 @@
-﻿//==================================================================================================================================//
-//!< @file		dx11Drawer.h
-//!< @brief		dx11::Drawerクラスヘッダ
-//!< @author	T.Haga
-//==================================================================================================================================//
-
-#ifndef SL_DX11_DRAWER_H
-#define SL_DX11_DRAWER_H
-
-namespace sl
-{
-namespace dx11
-{
-
-//======================================================================//
-//!< 描画を行うクラス
-//======================================================================//
-class Drawer
-{
-
-public:
-	/** Constructor */
-	Drawer(void);
-
-	/** Destructor */
-	~Drawer(void);
-};
-
-}	// namespace dx11
-}	// namespace sl
-
-#endif	// SL_DX11_DRAWER_H
-
-//==================================================================================================================================//
-// END OF FILE
-//==================================================================================================================================//
