+++ LibraryProject/SharokuLibrary/sl/sl.h
@@ -7,8 +7,8 @@
 #ifndef SL_H
 #define SL_H
 
-#include "Window\slIWindow.h"
-#include "Window\dx11Window\dx11Window.h"
+#include "Library/Window/dx11Window/dx11Window.h"
+
 #include "Graphics\dx11GraphicsDevice\dx11GraphicsDevice.h"
 #include "Texture\dx11Texture\dx11TextureManager.h"
 #include "Vertex\dx11Vertex\dx11Vertex2DManager.h"
