+++ LibraryProject/SharokuLibrary/sl/slBuild.h
@@ -12,6 +12,11 @@
 //!< Setting
 //======================================================================//
 
+/*  -------------------------------------------------------------------------------------------------------- */
+
+#define DIRECT_X_11 
+
+
 /* Screen ----------------------------------------------------------------------------------------------------- */
 
 #define ADJUST_CLIENT_SIZE					// クライアント領域を調整する(クライアント領域を指定したサイズに変換する場合)
