+++ LibraryProject/SharokuLibrary/sl/slISharokuLibrary.cpp
@@ -1,0 +1,46 @@
+//==================================================================================================================================//
+//!< @file		ISharokuLibrary.cpp
+//!< @brief		ISharokuLibraryƒNƒ‰ƒXŽÀ‘•
+//!< @author	T.Haga
+//==================================================================================================================================//
+
+/* Includes --------------------------------------------------------------------------------------------------- */
+
+#include "slISharokuLibrary.h"
+#include "slBuild.h"
+#include "Library/slDX11Library.h"
+#include "Common/slTemplate.h"
+
+/* Namespace -------------------------------------------------------------------------------------------------- */
+
+namespace sl
+{
+
+/* Static Variable -------------------------------------------------------------------------------------------- */
+
+ISharokuLibrary*		ISharokuLibrary::m_pILibrary = NULL;
+
+/* Public Functions ------------------------------------------------------------------------------------------- */
+
+void ISharokuLibrary::Create(void)
+{
+	if(m_pILibrary == NULL)
+	{
+
+#ifdef  DIRECT_X_11
+		m_pILibrary = new DX11Library();
+#endif
+
+	}
+}
+
+void ISharokuLibrary::Delete(void)
+{
+	DeleteSafly(m_pILibrary);
+}
+
+}	// namespace sl
+
+//==================================================================================================================================//
+// END OF FILE
+//==================================================================================================================================//
